webkit/data/layout_test_results/v8/chrome/fast/forms/textarea-metrics-expected.txt

   
This test checks that textareas have the right metrics.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.getElementById('textarea-no-scroll').clientWidth is 94
PASS document.getElementById('textarea-no-scroll').clientHeight is 94
PASS document.getElementById('textarea-no-scroll').offsetWidth is 96
PASS document.getElementById('textarea-no-scroll').offsetHeight is 96
PASS document.getElementById('textarea-no-scroll').scrollWidth is 94
PASS document.getElementById('textarea-no-scroll').scrollHeight is 20
FAIL document.getElementById('textarea-with-scroll').clientWidth should be 77. Was 79.
PASS document.getElementById('textarea-with-scroll').clientHeight is 94
PASS document.getElementById('textarea-with-scroll').offsetWidth is 96
PASS document.getElementById('textarea-with-scroll').offsetHeight is 96
FAIL document.getElementById('textarea-with-scroll').scrollWidth should be 77. Was 79.
PASS document.getElementById('textarea-with-scroll').scrollHeight is 132
PASS document.getElementById('textarea-no-scroll-with-padding').clientWidth is 106
PASS document.getElementById('textarea-no-scroll-with-padding').clientHeight is 106
PASS document.getElementById('textarea-no-scroll-with-padding').offsetWidth is 108
PASS document.getElementById('textarea-no-scroll-with-padding').offsetHeight is 108
PASS document.getElementById('textarea-no-scroll-with-padding').scrollWidth is 106
PASS document.getElementById('textarea-no-scroll-with-padding').scrollHeight is 32
FAIL document.getElementById('textarea-with-scroll-with-padding').clientWidth should be 89. Was 91.
PASS document.getElementById('textarea-with-scroll-with-padding').clientHeight is 106
PASS document.getElementById('textarea-with-scroll-with-padding').offsetWidth is 108
PASS document.getElementById('textarea-with-scroll-with-padding').offsetHeight is 108
FAIL document.getElementById('textarea-with-scroll-with-padding').scrollWidth should be 89. Was 91.
PASS document.getElementById('textarea-with-scroll-with-padding').scrollHeight is 144
